"ads2_m7_a1_1: 0/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/Module 7/m7/1. MST/Solution.java
@@ -11,7 +11,7 @@ class Solution {
 		while (sc.hasNextLine()) {
 			String[] tokens = sc.nextLine().split(" ");
 			Edge e = new Edge(Integer.parseInt(tokens[0]),
-			                  Integer.parseInt(tokens[1]), Integer.parseInt(tokens[2]));
+			                  Integer.parseInt(tokens[1]), Double.parseDouble(tokens[2]));
 			ewd.addEdge(e);
 		}
 		KruskalMST mst = new KruskalMST(ewd);
